+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Added
 
 - Watch videos only from allowed channels
+- Remove sidebar with recommended videos